Remove initial migration files
This commit removes the initial migration files for the 'FoodFood' project, as these files were no longer required for the current state of the database schema and operations. The deleted files include the migration.cs, migration.Designer.cs, and ModelSnapshot.cs files which originally handled database creation and schema versioning.