Add "RestaurantId" to Order model and enforce "UserId" as non-nullable
The changes enforce the "UserId" as a required field in the Order model, improving referential integrity. A new "RestaurantId" property was also added to the Order model to store the ID of the restaurant associated with the order. Test data for the Order table was removed from 'testData.sql' script as part of this update.